+++ tests/testthat/test-calculate_individual_radius_angle.R
@@ -1,11 +1,13 @@
 test_that(
-  "calc_animal_pos returns errors if animal_pos is not valid",
+  paste0("calculate_individual_radius_angle returns errors if ",
+         "animal_positions is not valid"
+  ),
   {
-    # animal_pos is not a dataframe
-    expect_error(calc_animal_pos(1, list(a = "a")))
-    # x, y and sequenceID columns missing
+    # animal_positions is not a dataframe
+    expect_error(calculate_individual_radius_angle(1, list(a = "a")))
+    # x, y and eventID columns missing
     expect_error(
-      calc_animal_pos(
+      calculate_individual_radius_angle(
         dplyr::tibble(
           deploymentID = "A",
           imageWidth = 5,
@@ -13,50 +15,54 @@ test_that(
         ),
         list(a = "a")
       ),
-      "Columns `sequenceID`, `x` and `y` not found in `animal_pos`."
+      "Columns `eventID`, `x` and `y` not found in `animal_positions`."
     )
     # imageWidth, imageHeight and deploymentID columns missing
     expect_error(
-      calc_animal_pos(
+      calculate_individual_radius_angle(
         dplyr::tibble(
           width = 5,
           height = 10,
-          sequenceID = 2,
+          eventID = 2,
           x = 4, y = 2
         ),
         list(a = "a")
       ),
-      "Columns `deploymentID`, `imageWidth` and `imageHeight` not found in `animal_pos`."
+      paste0("Columns `deploymentID`, `imageWidth` and `imageHeight` ",
+             "not found in `animal_positions`."
+      )
     )
   }
 )
 
 test_that(
-  "calc_animal_post returns errors if calib_models is not valid",
+  paste0("calculate_individual_radius_angle returns errors if ",
+         "calibration_models is not valid"
+  ),
   {
     df <- dplyr::tibble(
       imageWidth = 5,
       imageHeight = 10,
       deploymentID = "Z",
-      sequenceID = 3,
+      eventID = 3,
       x = 4,
       y = 2,
       deployment = 2
     )
-    # calib_models is not a (named) list
-    expect_error(calc_animal_pos(df, calib_models = 2))
+    # calibration_models is not a (named) list
+    expect_error(calculate_individual_radius_angle(df, calibration_models = 2))
     expect_error(
-      calc_animal_pos(df, calib_models = list(2)),
-      "`calib_models` must be a named list."
+      calculate_individual_radius_angle(df, calibration_models = list(2)),
+      "`calibration_models` must be a named list."
     )
   }
 )
 
 test_that("Deployments with no matching calibration model", {
-  missing_calib_model <- dep_calib_models
+  missing_calib_model <- calibration_models
   missing_calib_model$S01 <- NULL
   expect_warning(
-    calc_animal_pos(animal_positions, missing_calib_model),
+    calculate_individual_radius_angle(animal_positions, missing_calib_model),
     paste(
       "Some deployments have no matching calibration model",
       "and are stripped out: S01"
@@ -69,7 +75,7 @@ test_that("Deploys with multiple values for image width/height", {
   multi_pixel_dim$imageWidth[1] <- 4096
   multi_pixel_dim$imageHeight[20] <- 3072
   expect_warning(
-    calc_animal_pos(multi_pixel_dim, dep_calib_models),
+    calculate_individual_radius_angle(multi_pixel_dim, calibration_models),
     paste(
       "There is more than one unique value per deployment for `imageWidth`",
       "and/or `imageHeight` in deployment(s): S01, S02"
@@ -79,7 +85,9 @@ test_that("Deploys with multiple values for image width/height", {
 })
 
 test_that("Right output", {
-  output <- calc_animal_pos(animal_positions, dep_calib_models)
+  output <- calculate_individual_radius_angle(animal_positions,
+                                              calibration_models
+  )
   # right class
   expect_true(inherits(output, c("tbl_df", "tbl", "data.frame")))
   # right number of rows
@@ -99,20 +107,20 @@ test_that("Right output", {
 })
 
 test_that("Right output with non default column names", {
-  output_default <- calc_animal_pos(animal_positions, dep_calib_models)
+  output_default <- calculate_individual_radius_angle(animal_positions, calibration_models)
   animal_positions_non_default <- dplyr::rename(
     animal_positions,
     deployment_id = deploymentID,
-    sequence_id = sequenceID,
+    event_id = eventID,
     X = x,
     Y = y,
     image_width = imageWidth,
     image_height = imageHeight
   )
-  output <- calc_animal_pos(animal_positions_non_default,
-    dep_calib_models,
-    dep_tag = "deployment_id",
-    sequence_id = "sequence_id",
+  output <- calculate_individual_radius_angle(animal_positions_non_default,
+    calibration_models, 
+    deployment_id = "deployment_id",
+    event_id = "event_id",
     x = "X",
     y = "Y",
     image_width = "image_width",